requirements.txt
# Core RAG
llama-index==0.9.45
openai==1.12.0
python-dotenv==1.0.0

# Embeddings & Search
sentence-transformers==2.3.1
qdrant-client==1.7.0
rank-bm25==0.2.2

# Optimization
redis==5.0.1
numpy==1.24.3
pandas==2.0.3

# API & Frontend
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.31.0

# Document Processing
pypdf==4.0.1
python-multipart==0.0.9

# Testing & Benchmarking
pytest==8.0.0
pytest-cov==4.1.0
matplotlib==3.8.2
